ai_hackathon/Backend/requirements.txt
Django
djangorestframework
openai
python-dotenv
PyPDF2
tiktoken
numpy
